# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF640 - Service Activation and Configuration\n\
    \n### Release : 18.5 - December 2018\n\nService Activation and Configuration API\
    \ goal is to provide the ability to activate and configure Services.\n\n### Operations\n\
    Service Activation and Configuration API performs the following operations on\
    \ the resources :\n- Retrieve an entity or a collection of entities depending\
    \ on filter criteria\n- Partial update of an entity (including updating rules)\n\
    - Create an entity (including default values and creation rules)\n- Delete an\
    \ entity (for administration purposes)\n- Manage notification of events"
  version: "4.0.0"
  title: "API Service Activation and Configuration"
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
- "http"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
...
  /listener/monitorCreateNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity MonitorCreateNotification"
      description: "Example of a client listener for receiving the notification MonitorCreateNotification"
      operationId: "listenToMonitorCreateNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/MonitorCreateNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/monitorAttributeValueChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity MonitorAttributeValueChangeNotification"
      description: "Example of a client listener for receiving the notification MonitorAttributeValueChangeNotification"
      operationId: "listenToMonitorAttributeValueChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/MonitorAttributeValueChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/monitorStateChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity MonitorStateChangeNotification"
      description: "Example of a client listener for receiving the notification MonitorStateChangeNotification"
      operationId: "listenToMonitorStateChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/MonitorStateChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/monitorDeleteNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity MonitorDeleteNotification"
      description: "Example of a client listener for receiving the notification MonitorDeleteNotification"
      operationId: "listenToMonitorDeleteNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/MonitorDeleteNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceCreateNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceCreateNotification"
      description: "Example of a client listener for receiving the notification ServiceCreateNotification"
      operationId: "listenToServiceCreateNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceCreateNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceAttributeValueChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceAttributeValueChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceAttributeValueChangeNotification"
      operationId: "listenToServiceAttributeValueChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceAttributeValueChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceStateChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceStateChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceStateChangeNotification"
      operationId: "listenToServiceStateChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceStateChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceDeleteNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceDeleteNotification"
      description: "Example of a client listener for receiving the notification ServiceDeleteNotification"
      operationId: "listenToServiceDeleteNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceDeleteNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
definitions:
...
  EventSubscription:
    type: "object"
    required:
    - "callback"
    - "id"
    properties:
      id:
        type: "string"
        description: "Id of the listener"
      callback:
        type: "string"
        description: "The callback being registered."
      query:
        type: "string"
        description: "additional data to be passed"
    description: "Sets the communication endpoint address the service instance must\
      \ use to deliver notification information"
    example:
      query: "query"
      callback: "callback"
      id: "id"
...
  MonitorCreateNotification:
    type: "object"
    properties:
      eventId:
        type: "string"
        description: "The identifier of the notification"
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurence"
      eventType:
        type: "string"
        description: "The type of the notification"
      fieldPath:
        type: "string"
        description: "The path identifying the object field concerned by this notification"
      resourcePath:
        type: "string"
        description: "The path identifying the resource object concerned by this notification"
      event:
        description: "The event linked to the involved resource object"
        $ref: "#/definitions/MonitorCreateEvent"
    description: "The notification data structure"
    example:
      eventId: "eventId"
      resourcePath: "resourcePath"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      fieldPath: "fieldPath"
      eventType: "eventType"
      event:
        monitor:
          request:
            method: "method"
            '@baseType': "@baseType"
            '@type': "@type"
            header:
            - '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: "value"
            - '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: "value"
            to: "to"
            body: "body"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          response:
            '@baseType': "@baseType"
            '@type': "@type"
            header:
            - '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: "value"
            - '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              value: "value"
            body: "body"
            '@schemaLocation': "http://example.com/aeiou"
            statusCode: "statusCode"
          '@type': "@type"
          id: "id"
          href: "href"
          state: "state"
          '@schemaLocation': "http://example.com/aeiou"
          sourceHref: "sourceHref"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"